Temporarily replace built-in webcrypto with @peculiar/webcrypto
The webcrypto API was introduced in Node 15.
AWS Lambda (and by extension, Vercel and Netlify) only supports LTS versions of Node.
Therefore, we need to use an alternative until Node 16 is supported.

I tried replacing it with the built-in crypto module in Node, but I'm afraid it would be incorrect.