README: misleading lock-in statement
The documentation is wildly misleading. Everything in this project gets
locked behind a for-profit, closed source, proprietary service in
Microsoft™ GitHub™—from who gets to keep the data, to account
creation/deletion. It’s misleading to say that this isn’t tracked & has
no lock-in, when it clearly does. We would be better off leveraging the
*decentralized* part of distributed version control systems (DVCS)
instead of centralizing around a service like Microsoft™ GitHub™.